pyproject.toml
[tool.poetry]
name = "file_processing"
version = "0.1.0"
description = ""
authors = ["Dobro <matematikaigor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.3"
python-dotenv = "^0.20.0"
python-dateutil = "^2.8.2"
pydantic = "^1.9.2"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
autopep8 = "^1.5.0"
flake8 = "^4.0.1"
wemake-python-styleguide = "^0.16.1"
black = "^22.6.0"
pylint = "^2.14.5"
pytest-cov = "^4.0.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
